pyproject.toml
[project]
name = "hexagone-2b-2insi-2026-tp3"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.80.0",
    "grpcio-tools>=1.80.0",
    "psycopg2-binary>=2.9.11",
    "requests>=2.33.0",
    "sqlalchemy>=2.0.48",
]
